[EPROD-697] Add i686 CI tests

+++ ic-stable-structures/src/memory_mapped_files/error.rs
@@ -5,7 +5,7 @@ pub enum MemMapError {
     #[error("file error: {0}")]
     FileOpenError(#[from] std::io::Error),
     #[error("address space limit exceeded")]
-    OutOfAddressSpace { claimed: u64, limit: u64 },
+    OutOfAddressSpace { claimed: usize, limit: usize },
     #[error("access out of bounds")]
     AccessOutOfBounds,
     #[error("new length should be page size multiple")]
